requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Airtable
pyairtable==2.3.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# SendGrid SDK
sendgrid==6.11.0

# Utils
httpx==0.25.2
python-dateutil==2.8.2
requests==2.32.3 